+++ ShoppingCart/ShoppingCart/Implementation/ShippingCalculator.cs
@@ -5,9 +5,9 @@ namespace ShoppingCart.Implementation
     public class ShippingCalculator : IShippingCalculator
     {
         // TODO this should come from config or be provided by a ShippingThresholdRepository
-        private (double value, double price)[] _shippingThresholds = new[] { (20.0, 7.0), (40.0, 5.0) };
+        private (decimal value, decimal price)[] _shippingThresholds = new[] { (20.0m, 7.0m), (40.0m, 5.0m) };
 
-        public double CalcShipping(double cartTotal)
+        public decimal CalcShipping(decimal cartTotal)
         {
             foreach (var threshold in _shippingThresholds)
                 if (cartTotal < threshold.value)
