+++ ShoppingCart/ShoppingCart/Models/Product.cs
@@ -7,7 +7,7 @@
         public string Name { get; set; }
         public string[] Categories { get; set; }
         public string Supplier { get; set; }
-        public double WholesalePrice { get; set; }
-        public double Price { get; set; }
+        public decimal WholesalePrice { get; set; }
+        public decimal Price { get; set; }
     }
 }
